Fix pylint errors.
