+++ futaba/navi/punish.py
@@ -10,7 +10,6 @@
 # WITHOUT ANY WARRANTY. See the LICENSE file for more details.
 #
 
-import asyncio
 import logging
 
 import discord
