+++ futaba/punishment.py
@@ -14,10 +14,7 @@
 Contains commands for handling punishment roles in a modular way.
 """
 
-import asyncio
 import logging
-
-import discord
 
 logger = logging.getLogger(__name__)
 
@@ -31,52 +28,43 @@ class PunishmentHandler:
     def __init__(self, bot):
         self.bot = bot
 
-    def check_other_roles(self, member):
-        has_other, punish_role, _ = self.bot.sql.moderation.get_other_roles(member)
-        if has_other:
-            embed = discord.Embed(colour=discord.Colour.red())
-            role_descr = (
-                ""
-                if punish_role is None
-                else f"because they already have {punish_role.mention}"
-            )
-            embed.description = (
-                f"Cannot add a new overriding role to {member.mention} {role_descr}"
-            )
-            raise CommandFailed(embed=embed)
-
-    def get_role(self, name):
-        roles = self.bot.sql.settings.get_special_roles(ctx.guild)
+    def get_role(self, name, guild, member):
+        roles = self.bot.sql.settings.get_special_roles(guild)
 
         role = getattr(roles, name)
         if role is None:
-            logger.error("No %s role configured for guild '%s' (%d)", name, guild.name, guild.id)
+            logger.warning("No %s role configured for guild '%s' (%d)", name, guild.name, guild.id)
+            return None
+
+        has_other, _, _ = self.bot.sql.moderation.get_other_roles(member)
+        if has_other:
+            logger.warning("Cannot add an overriding role to '%s' (%d)", member.name, member.id)
+            return None
 
         return role
 
-    async def apply(self, action, guild, member, reason):
-        role = self.get_role(action)
+    async def apply(self, name, guild, member, reason):
+        role = self.get_role(name, guild, member)
         if role is None:
             return
 
         if member.top_role > guild.me.top_role:
-            logger.error("Lacks permission to %s user '%s' (%d) in guild '%s' (%d)", action, member.name, member.id, guild.name, guild.id)
+            logger.warning("Lacks permission to %s user '%s' (%d) in guild '%s' (%d)", name, member.name, member.id, guild.name, guild.id)
             return
 
         remove_other = self.bot.sql.settings.get_remove_other_roles(guild)
         if remove_other:
-            self.check_other_roles(member)
             await self.bot.sql.moderation.remove_other_roles(member, role, reason)
         else:
             await member.add_roles(role, reason=reason)
 
-    async def relieve(self, action, guild, member, reason):
-        role = self.get_role('mute')
+    async def relieve(self, name, guild, member, reason):
+        role = self.get_role(name, guild, member)
         if role is None:
             return
 
         if member.top_role > guild.me.top_role:
-            logger.error("Lacks permission to unmute user '%s' (%d) in guild '%s' (%d)", member.name, member.id, guild.name, guild.id)
+            logger.warning("Lacks permission to %s user '%s' (%d) in guild '%s' (%d)", name, member.name, member.id, guild.name, guild.id)
             return
 
         remove_other = self.bot.sql.settings.get_remove_other_roles(guild)
@@ -84,7 +72,7 @@ class PunishmentHandler:
             try:
                 await self.bot.sql.moderation.restore_other_roles(member, reason)
             except KeyError as error:
-                logger.warning(f"Received KeyError while restoring other roles: {error}")
+                logger.warning("Received KeyError while restoring other roles: %s", error)
 
         await member.remove_roles(role, reason=reason)
 
@@ -103,6 +91,3 @@ class PunishmentHandler:
     async def unjail(self, guild, member, reason = None):
         logger.info("Unjailing user '%s' (%d) for reason: %s", member.name, member.id, reason)
         await self.relieve('jail', guild, member, reason)
-
-    async def unmute(self, guild, member, reason = None):
-        logger.info("Unmuting user '%s' (%d)", member.name, member.id)
